🚚 [Move]: #472 task-links を「1 型名 = 1 フォルダ（index.ts + __tests__）」構成へ再編
責務ファイル分割（plan.ts / linkOperation.ts 等）を、型名単位のフォルダ構成に改める:

- TaskLinks/       TaskLinks 型 + companion（planAddLink / planRemoveLink・
                   candidates・removeLinkedTask）+ TaskLinks.*.test.ts 7 件
- LinkOperation/   LinkOperation 型 + 適用・グルーピング・inverse 変換
- LinkIntent/      LinkIntent 型 + companion（forAdd / forRemove）+ factory テスト
- index.ts         re-export のみ（公開 API 不変）

plan の計算は aggregate method として TaskLinks companion に同居させ、
operations の変換（invertLinkOperations）は LinkOperation 側に置く。
テストは各型フォルダの __tests__ へ移動し、相対 import を @/ alias に統一。
ロジック無変更・全テスト無改修 PASS。

Refs #472
