--- src/domains/task-links/types.ts
@@ -1,7 +1,0 @@
-/** Task の関連リンク情報 */
-export type TaskLinks = {
-  /** 関連タスクのファイルパスの配列 */
-  linkedFilePaths: string[];
-  /** 逆方向リンクのファイルパスの配列（links から逆引き） */
-  reverseLinkedFilePaths: string[];
-};
